+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### New features
-`core` module
+`core-mit` module
 - Added `eraseStackTrace()` method in `EncodableResult` & `WrappedException` which erases the whole stack trace in exception information
+- Added `randomBase64()` method to generate random Base64 of the specified length
 ### Fixed Bugs
 `reflection` module:
 - `.encodeObject()`/`.decodeObject()` not working properly with `null`